pyproject.toml
[tool.poetry]
name = "airflow-tutorial"
version = "0.1.0"
description = ""
authors = ["Hank Ehly <henry.ehly@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
apache-airflow = "^1.10.9"

[tool.poetry.dev-dependencies]
isort = "^4.3.21"
black = "^19.10b0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.isort]
multi_line_output=3
include_trailing_comma=true
force_grid_wrap=0
use_parentheses=true
line_length=88
